### 2024-08-29 #7159 @espottesmith
Possible reviewers:

- [Anshuman Kumar](https://reviewers.joss.theoj.org/reviewers/3238)
- [@lorenzo-rovigatti](https://reviewers.joss.theoj.org/reviewers/1538)
- [@ptmerz](https://reviewers.joss.theoj.org/reviewers/1919)
- [@corettialessandro](https://reviewers.joss.theoj.org/reviewers/1387)
