### 2024-09-16 #7159 @mbarzegary
Hi @Anshuman5, @lorenzo-rovigatti, @ptmerz, @corettialessandro 👋
Considering your field of expertise as well as your experience with JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.07159/joss.07159/10.21105.joss.07159.pdf
Software: https://github.com/BlauGroup/RNMC
