### 2024-09-18 #7244 @editorialbot
[REVIEW]: RNMC: kinetic Monte Carlo implementations for complex reaction networks
**Submitting author:** @espottesmith (<a href="http://orcid.org/0000-0003-1554-197X">Evan Spotte-Smith</a>)
**Repository:** https://github.com/BlauGroup/RNMC
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
