### 2024-09-18 #7244 @mbarzegary
@espottesmith this is where the review takes place. Please keep an eye out for comments here from the reviewers, as well as any issues opened by them on your software repository. I recommend you aim to respond to these as soon as possible, and you can address them straight away as they come in if you like, to ensure we do not loose track of the reviewers.

To start, can you please fix the DOIs issue raised by the editorial bot above?
