### 2024-09-19 b0e240a
Update paper w/ DOIs
