+++ paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={679--701},
   year={1965},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1696792}
 }
 
 @article{gillespie1977exact,
@@ -17,7 +18,7 @@
   number={25},
   pages={2340--2361},
   year={1977},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
 }
 
 @techreport{garcia2009crossing,
@@ -35,7 +36,8 @@
   number={7},
   pages={2138--2150},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2014.04.003}
 }
 
 @book{newman2021electrochemical,
@@ -53,7 +55,8 @@
   number={4},
   pages={1446--1453},
   year={2022},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acsenergylett.2c00517}
 }
 
 @article{barter2023predictive,
@@ -64,18 +67,20 @@
   number={1},
   pages={123--137},
   year={2023},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi={10.1039/D2DD00117A}
 }
 
 @article{spotte2023chemical,
-  title={Chemical reaction networks explain gas evolution mechanisms in mg-ion batteries},
+  title={Chemical reaction networks explain gas evolution mechanisms in {Mg}-ion batteries},
   author={Spotte-Smith, Evan Walter Clark and Blau, Samuel M and Barter, Daniel and Leon, Noel J and Hahn, Nathan T and Redkar, Nikita S and Zavadil, Kevin R and Liao, Chen and Persson, Kristin A},
   journal={Journal of the American Chemical Society},
   volume={145},
   number={22},
   pages={12181--12192},
   year={2023},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/jacs.3c02222}
 }
 
 @article{xia2023accelerating,
@@ -86,7 +91,8 @@
   number={23},
   pages={11129--11136},
   year={2023},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.nanolett.3c03568}
 }
 
 @article{chan2015combinatorial,
@@ -97,7 +103,8 @@
   number={6},
   pages={1653--1679},
   year={2015},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi={10.1039/C4CS00205A}
 }
 
 @article{skripka2023NL,
@@ -109,8 +116,6 @@
   pages = {7100-7106},
   year = {2023},
   doi = {10.1021/acs.nanolett.3c01955},
-  URL = {https://doi.org/10.1021/acs.nanolett.3c01955},
-  eprint = {https://doi.org/10.1021/acs.nanolett.3c01955},
 }
 
 @article{teitelboim2019energy,
@@ -121,5 +126,6 @@
   number={4},
   pages={2678--2689},
   year={2019},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.jpcc.9b00161}
 }
