### 2024-09-22 BlauGroup/RNMC#12 @lorenzo-rovigatti
[JOSS review] Installation issues
Hi! This is an issue that's part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7244)

I'm trying to build the software on my computer, which runs Ubuntu 22.04. Unfortunately, linking fails with a bunch of `undefined reference to sqlite3_[...]` errors. Note that the packages `sqlite3`, `libsqlite3-dev` and `libsqlite3-0` are all installed. Can you help me?

More in general, I think that it would be nice if you could devise a better way of installing the software than just makefiles. My advice is to consider using cmake (or even the venerable configure) to manage the pre-compilation and compilation stages.
