### 2024-09-22 BlauGroup/RNMC#13 @lorenzo-rovigatti
[JOSS review] Some information missing
Hi! This is an issue that's linked to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7244)

According to the JOSS guidelines, there should be clear guidelines for third parties wishing to

1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

As far as I can see these are all missing in the current version of the repo. My advice is to set up [issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) for n. 2, and to add to the README and/or to the docs some information about what people should do to contribute to your software.
