### 2024-09-22 #7244 @lorenzo-rovigatti
I have added a couple of issues to the repo. For now I can't install the software and therefore I can't test it, but I have read the paper and I found it good, apart from the following two issues:

* Authorship: the list of authors is rather long compared to [the people who are listed as contributors](https://github.com/BlauGroup/RNMC/graphs/contributors). Do all authors comply with the [JOSS policies](https://joss.readthedocs.io/en/latest/submitting.html#authorship) on authorship?
* Software extensibility: the paper states that `RNMC` has been designed "to be easily extensible". Can you clarify this statement in the paper and perhaps in the documentation, which lacks a section on this matter?
*
