### 2024-09-24 1f02d29
Working on adding a guide to contributors
