+++ contributors.md
@@ -34,18 +34,17 @@ $ git remote add upstream https://github.com/BlauGroup/RNMC
 $ git push --set-upstream origin <BRANCH_NAME> 
 ```
 
-4. When you're ready, create a [pull request](https://github.com/BlauGroup/RNMC/pulls) on GitHub. “Work-in-progress” pull requests are encouraged, especially if this is your first time contributing to `RNMC`. If your feature is not finished, put "\[WIP\]" in the pull request title.
+4. When you're ready, create a [pull request](https://github.com/BlauGroup/RNMC/pulls) on GitHub. “Work-in-progress” pull requests are encouraged, especially if this is your first time contributing to `RNMC`. If your feature is not finished, or you anticipate needing significant feedback/revisions before the PR is merged in, put "\[WIP\]" in the pull request title.
 
 ## Raising Issues
 
-If you have a problem installing or using `RNMC`, you can raise an issue on our [GitHub page](https://github.com/BlauGroup/RNMC/issues). The issues page is also an appropriate place to ask questions or to suggest a new feature. If you create an issue, you can label it to help us triage appropriately.
+If you have a problem installing or using `RNMC`, you can raise an issue on our [GitHub page](https://github.com/BlauGroup/RNMC/issues). The issues page is also an appropriate place to ask questions or to suggest a new feature. If you create an issue, you can label it (e.g., "bug", "enhancement", or "question") to help us triage appropriately.
 
 ## Your First Contribution
 
 If you have a feature that you want to add but don't know where to start: reach out to [Sam Blau](mailto:smblau@lbl.gov) and describe what you have in mind. He, or one of the developers, may be able to help you get started.
 
-<!-- TODO: link to the page for expanding RNMC -->
-If you're specifically interested in adding a new type of simulator to RNMC: check back soon! We're working on a guide to expanding `RNMC`.
+If you're specifically interested in adding a new type of simulator to RNMC: check back our guide to [expanding RNMC](./Expanding.md), which details the basic components that you'll need to create a simulator.
 
 If you want to help but don't have a particular idea in mind, check out the issues page and see if there's something that you would be interested in contributing.
 
@@ -57,11 +56,13 @@ If you want to help but don't have a particular idea in mind, check out the issu
 
 ### Dependencies
 
-`RNMC` is a slim codebase, with `sqlite3` and `GSL` being the only dependencies. In general, we encourage you to only use the C/C++ standard library and these existing dependencies and not to add any external dependencies. If you believe that an external dependency is necessary, please discuss this with the `RNMC` developers and maintainers.
+`RNMC` is a relatively slim codebase with only two external dependencies: `sqlite3` and `GSL`. We encourage you to only use the C/C++ standard library and these existing dependencies and to avoid adding additional dependencies. If you believe that an external dependency is necessary, please discuss this with the `RNMC` developers and maintainers.
 
 ### Documentation
 
-TODO
+We do not require documentation on a per-file or per-function basis but encourage comments that explain the steps involved in different processes. If you add new features, we may request additional code documentation to assist with maintainability.
+
+New features (and especially new simulators!) should be described in the repository README and in the documentation hosted on [GitHub Pages](https://blaugroup.github.io/RNMC/). Note that information for GitHub pages is located in the `gh-pages` branch of the `RNMC` repository, so any PRs to add documentation to our project website should modify and merge into `gh-pages`, rather than `main`.
 
 ### Testing
 
