### 2024-09-26 cfb28d1
First draft on expanding RNMC
