### 2024-09-26 1bd2683
Add contributors page to the sidebar menu
