### 2024-09-30 BlauGroup/RNMC#13 @ptmerz
Other reviewer here. I'll note that https://blaugroup.github.io/RNMC/ does have a section which covers the three points listed above. I think this is technically sufficient to pass the JOSS requirement, but I agree with @lorenzo-rovigatti that having these things mirrored on the repo README would be valuable!
