### 2024-09-30 BlauGroup/RNMC#15 @ptmerz
[JOSS review] Installation of single module requires previous folder creation
Minor thing, but took me way too long to understand what was going wrong:

In https://blaugroup.github.io/RNMC/Installation.html, you mention that one could build single modules by `cd`-ing into the respective folder and calling `make MODULE` there. This throws a non-descriptive linker error message, _unless_ a directory called `build` was previously created.

It would be great if you could create this folder within your makefile. Otherwise, please at least assert for whether the folder is there and fail with a more informative error message.
