### 2024-09-30 #7244 @ptmerz
First part of my review, unfortunately I ran out of time today and will continue in the next days.

* License: While the license looks like an appropriate open source license, I didn't succeed on finding it on https://opensource.org/licenses/alphabetical as requested by the checklist. If it is on the list, can you point me to it?
* Contributors: As @lorenzo-rovigatti mentioned, we can only verify three of the authors (Laura Zichi, Daniel Barter, and @espottesmith) from the repo. Can you please comment on your choices for the author list?
* Installation: While I could install the software without issues in one of the described ways, the other described way failed with a non-descriptive error. The underlying problem is very minor, but the resulting error can be confusing and time consuming. I opened an issue at https://github.com/BlauGroup/RNMC/issues/15.
