### 2024-09-30 #7244 @ptmerz
And a quick note on the references in the paper: Some of the references in the text are rendered a bit weirdly. Several references either lack a space between the previous word and the parenthesis like this
> Marcus theory(Marcus, 1965)
or they are rendered after the period or comma rather than before:
> Butler-Volmer kinetics.(Newman & Balsara, 2021)

A quick look at recent publications (https://joss.theoj.org) confirm that they should be rendered with a space and before any period or comma.
