### 2024-10-04 BlauGroup/RNMC#13 @espottesmith
Good suggestion! We'll provide a briefer description of these points in the README
