### 2024-10-04 #7244 @espottesmith
Hey, I'll be addressing these questions/comments piecemeal.

First, references should be fixed.

Regarding the license, we're using the BSD-3-Clause-LBNL license. From [https://spdx.org/licenses/BSD-3-Clause-LBNL.html](here):
> This license is the same as BSD-3-Clause, but with an additional default contribution licensing clause
