### 2024-10-07 BlauGroup/RNMC#12 @espottesmith
Hey Lorenzo,

Thank you for raising this issue. We're trying to diagnose what's going on. For us to reproduce the issue, it'd be helpful if you could give us some more information. Specifically, the following would be helpful:

1. Full error messages
2. Your `LIBRARY_PATH` and/or `LD_LIBRARY_PATH`.

We're also looking into changing our build process to use `cmake`, though, if we can resolve this issue and make our makefiles more robust, we may keep with our current installation method.
