### 2024-10-07 b6f7281
Expanded README to add in guide for contributors
