### 2024-10-07 #7244 @espottesmith
Regarding authorship:

- Eric Sivonxay (@sivonxay) wrote most of the `NPMC` simulator for nanoparticle photophysics. He opened a PR (https://github.com/BlauGroup/RNMC/pull/3), but we ultimately decided that it'd be easier if @lzichi just took his code and did the work of unifying the various simulators.
- Rohith Srinivaas Mohanakrishnan (@rohithsrinivaas) did a lot of testing and contributed to bugfixes. He was particularly helpful for testing and developing `LGMC`. He also worked on some code for visualization and analysis, but this code hasn't been integrated into `RNMC` yet.
- Emory M. Chan (not sure he's on GitHub?) provided code for nanoparticle simulations to Eric, which eventually became `NPMC`. He also provided direct technical guidance for the development of `NPMC`.
- Kristin A. Persson (also not on GitHub?) provided technical mentorship to several members of this team, including myself, Laura Zichi, and Rohith Srinivaas
- Samuel M. Blau (@samblau) conceived of the idea of `RNMC`, directed its development from inception, and provided technical guidance to most of the developers. I'm sure he also did some amount of testing, code review, etc., though @danielbarter might be able to speak to that with more detail.
