### 2024-10-08 BlauGroup/RNMC#12 @lorenzo-rovigatti
Sure thing! Here is the output of `build.sh`: [errors.txt](https://github.com/user-attachments/files/17293330/errors.txt)
Both env variables are empty. Note that the following program compiles with `g++ -o sql sql.cpp -lsqlite3`:

```C++
#include <iostream>
#include <sqlite3.h>

int main(int argc, char** argv)
{
    sqlite3* DB;
    int exit = 0;
    exit = sqlite3_open("example.db", &DB);

    if (exit) {
        std::cerr << "Error open DB " << sqlite3_errmsg(DB) << std::endl;
        return (-1);
    }
    else
        std::cout << "Opened Database Successfully!" << std::endl;
    sqlite3_close(DB);
    return (0);
}
```
