### 2024-10-08 BlauGroup/RNMC#13 @lorenzo-rovigatti
I think that's enough to comply with JOSS's requirements. I close this one, thank you!
