### 2024-10-09 43ee733
fixed linker and build dir errors
