### 2024-10-09 BlauGroup/RNMC#12 @espottesmith
Thanks for that detail! I believe this should have been fixed by https://github.com/BlauGroup/RNMC/pull/16. Please check and see if you can install now.
