### 2024-10-09 #7244 @espottesmith
Installation issues should now be fixed, thanks to @lzichi.

To @lorenzo-rovigatti's point about extending RNMC, there's also now a page in the documentation on [expanding RNMC](https://blaugroup.github.io/RNMC/Expanding.html), which is linked in the page for [contributors](https://blaugroup.github.io/RNMC/Contributors.html). I can add a brief section in the manuscript as well.
