### 2024-10-10 BlauGroup/RNMC#12 @lorenzo-rovigatti
It works now, thank you!
