### 2024-10-24 #7244 @Anshuman5
This work presents a highly valuable tool for the scientific community by providing a program for kMC simulation for modeling complex systems. I strongly recommend publication after addressing the following:
- include current limitation of RNMC - for example, limited to simulating reactions upto two reactants and two products, in the draft or readme.md file.
- in the documentation under GMC - Running GMC, include the sample output file names, logs, and directories that will be generated after running the command `build/GMC --reaction_database=examples/GMC/end-to-end-test/...`. Do the same for other sections - NPMC and LGMC.
- RNMC is not expanded in the Summary section of the draft (line no.21).
- Change the color of text 'RNMC' in logo.png. It is not clearly visible when opening github in dark mode. See attached figure.
![image](https://github.com/user-attachments/assets/ecc655b3-7487-48e7-be43-d1dcffa7ddd4)
