### 2024-10-24 246f586
Change logo text color and make small paper and README changes in response to reviewer comments
