+++ LGMC.md
@@ -183,3 +183,5 @@ When running LGMC ensure that your input file paths are correct considering the 
 ```
 build/LGMC --lattice_reaction_database=examples/LGMC/CO_oxidation/rn.sqlite --initial_state_database=examples/LGMC/CO_oxidation/initial_state.sqlite --number_of_simulations=1000 --base_seed=1000 --thread_count=8 --step_cutoff=200 --checkpoint=1 --parameters=examples/LGMC/CO_oxidation/LGMC_params.txt
 ```
+
+Running this command does not generate any new files or directories but will populate the `initial_state_database` with trajectory data.
