### 2024-10-25 #7244 @espottesmith
Thank you for your comments, @Anshuman5. I've now addressed all of them. I added a "limitations" section to the README, included the full name of `RNMC` in the paper draft, and made the logo gray rather than black for dark mode users.

To your point about file/directory generation: no new files/directories are generated by running `GMC`/`NPMC`/`LGMC`. All data is merely inserted into an existing database file. I have now clarified this in the docs.
