### 2024-11-19 #7244 @ptmerz
I finished my review checklist and would strongly recommend the paper for acceptance in JOSS. The software, documentation and repo are currently in excellent shape, and the paper is well written. Thank you to the authors for addressing the issues brought up during the review!
