### 2024-11-25 #7244 @mbarzegary
@espottesmith given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with some minor edits
* Work on the author's points of the final checklist I created ☝️
