### 2024-11-25 #7244 @espottesmith
- Authors and affiliations are correct
- The most recent release, which includes all changes related to this review, is https://github.com/BlauGroup/RNMC/releases/tag/v1.1.0
- There's currently an issue with the RNMC Zenodo (https://zenodo.org/records/7277767), and recent releases haven't been posted and archived there. I'll work to get that sorted as soon as possible. Once I am able to post releases there, I'll also update the author list and license.
