### 2024-12-10 #7244 @espottesmith
Hey @mbarzegary, sorry for the delay.

I wasn't able to get v1.1.0 onto Zenodo because of some permission issues, but I was able to release a new version and get [v1.1.1](https://github.com/BlauGroup/RNMC/releases/tag/v1.1.1) archived. The DOI is 10.5281/zenodo.14360064.

I've updated the title and author list to match the paper, and the licenses match.

Please let me know if there's anything else you need to move this process forward!
