### 2024-12-13 #7244 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
