### 2024-12-17 #7244 @Kevin-Mattheus-Moerman
@espottesmith as AEiC for JOSS I will now help to process this submission for acceptance in JOSS.

:point_right: seems to not be recognised by GitHub but is the bsd-3-clause-lbnl license: https://opensource.org/license/bsd-3-clause-lbnl
